back-end/requirements.txt
asgiref==3.8.1
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
platformdirs==4.2.1
psycopg==3.1.18
psycopg-binary==3.1.18
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.2.1
